move the ori-wrapped cluster logic at the very end of find_clusters subroutine, simplify it a bit
If i understood correctly, at the "borders" prediction stage, there can be no "Compound1-CompoundX" borders, right? This happens only later, in the "clusters" stage? If so, this allows to simplify this subroutine a bit, which i made too complex before. Also, shoudn't we rename the subroutine to "find_borders"?