secmet: remove the location_bridges_origin() and split/bridge test leftovers
in my view, it might be impossible to reliably determine the outcome of this subroutine, let alone testing it
# consider the two forward-strand features: CompoundLocation([FeatureLocation(990,1000),FeatureLocation(0,10)])
and
CompoundLocation([FeatureLocation(500,510),FeatureLocation(495,505)])
on a circular record. How one has to determine which one is bridging, and which one is just weird exon order? Even if we knew the len(record), it would be still a hard case to determine which is which. Not without guessing that is.
Instead of this, we would like to suggest a more generalized way of dealing with cases, mostly by extensively checking the overlaps of subfeatures while growing cores/clusters/superclusters/regions.