hmm_detection: Add rule detecting lipolanthines
Of the sequences mentioned in https://doi.org/10.1038/s41589-018-0068-6,
This rule detects all but three of the clusters.

The clusters in NZ_BDAY01000007.1 and NZ_JNZO01000011.1 are missed due
to bad ORF calls in the biosynthesis enzyme. It's unclear what is going
on with NZ_JOAG00000000, as antiSMASH doesn't even find a Flavoprotein
hit there.

Signed-off-by: Kai Blin <kblin@biosustain.dtu.dk>