common.record_processing: handle bug in macOS py3.8+ pickling config
Beginning in python 3.8 on macOS `multiprocessing.Pool` started using
`spawn()` instead of `fork()`, which broke the singleton behaviour
with `get_config()`. This only affected `record_processing` calling
the gene finding detection module.

This fixes this specific issue by reconstructing the relevant options
within the function needing them in the case that that the config
object is empty.